Refactor price parser, a downgrade unknown card names to warnings.
