+++ price_grabber/price_grabber.py
@@ -7,16 +7,13 @@ import ftfy
 
 from magic import card, fetcher_internal, multiverse, oracle
 from magic.database import db
-from price_grabber import price
+from price_grabber import price, parser
 from shared import configuration, dtutil
 from shared.database import get_database
 from shared.pd_exception import (DatabaseException, InvalidDataException,
                                  TooFewItemsException)
 
 DATABASE = get_database(configuration.get_str('prices_database'))
-CARDS: Dict[str, str] = {}
-
-PriceList = List[Tuple[str, str, str]] # pylint: disable=invalid-name
 
 def run() -> None:
     multiverse.init()
@@ -30,46 +27,18 @@ def fetch() -> None:
         s = fetcher_internal.fetch(url)
         s = ftfy.fix_encoding(s)
         timestamps.append(dtutil.parse_to_ts(s.split('\n', 1)[0].replace('UPDATED ', ''), '%Y-%m-%dT%H:%M:%S+00:00', dtutil.CARDHOARDER_TZ))
-        all_prices[url] = parse_cardhoarder_prices(s)
+        all_prices[url] = parser.parse_cardhoarder_prices(s)
     url = configuration.get_str('mtgotraders_url')
     if url:
         s = fetcher_internal.fetch(url)
         timestamps.append(dtutil.dt2ts(dtutil.now()))
-        all_prices['mtgotraders'] = parse_mtgotraders_prices(s)
+        all_prices['mtgotraders'] = parser.parse_mtgotraders_prices(s)
     if not timestamps:
         raise TooFewItemsException('Did not get any prices when fetching {urls} ({all_prices})'.format(urls=configuration.get_list('cardhoarder_urls') + [configuration.get_str('mtgotraders_url')], all_prices=all_prices))
     store(min(timestamps), all_prices)
 
-def parse_cardhoarder_prices(s: str) -> PriceList:
-    details = []
-    for line in s.splitlines()[2:]: # Skipping date and header line.
-        if line.count('\t') != 6:
-            raise InvalidDataException('Bad line (cardhoarder): {line}'.format(line=line))
-        else:
-            _mtgo_id, mtgo_set, _mtgjson_set, set_number, name, p, quantity = line.split('\t')  # pylint: disable=unused-variable
-            name = html.unescape(name.strip())
-            if int(quantity) > 0 and not mtgo_set.startswith('CH-') and mtgo_set != 'VAN' and mtgo_set != 'EVENT' and not re.search(r'(Booster|Commander Deck|Commander:|Theme Deck|Draft Pack|Duel Decks|Reward Pack|Intro Pack|Tournament Pack|Premium Deck Series:|From the Vault)', name):
-                details.append((name, p, mtgo_set))
-    return [(name_lookup(name), html.unescape(p.strip()), mtgo_set) for name, p, mtgo_set in details if name_lookup(name) is not None]
 
-def parse_mtgotraders_prices(s: str) -> PriceList:
-    details = []
-    for line in s.splitlines():
-        if line.count('|') != 7:
-            raise InvalidDataException('Bad line (mtgotraders): {line}'.format(line=line))
-        mtgo_set, rarity, premium, name, number, p, image_path, in_stock_str = line.split('|') # pylint: disable=unused-variable
-        in_stock_str = in_stock_str.replace('<br>', '')
-        assert in_stock_str == 'Yes' or in_stock_str == 'No'
-        assert p == '0.01' or p == '0.00'
-        in_stock = in_stock_str == 'Yes'
-        if p == '0.01' and in_stock and not is_exceptional_name(name):
-            details.append((name, p, mtgo_set))
-    return [(name_lookup(name), p, mtgo_set) for name, p, mtgo_set in details if name_lookup(name) is not None]
-
-def is_exceptional_name(name: str) -> bool:
-    return name.startswith('APAC ') or 'Alternate art' in name or name.startswith('Avatar - ') or name.startswith('Euro ')
-
-def store(timestamp: float, all_prices: Dict[str, PriceList]) -> None:
+def store(timestamp: float, all_prices: Dict[str, parser.PriceList]) -> None:
     DATABASE.begin()
     lows: Dict[str, int] = {}
     for code in all_prices:
@@ -126,16 +95,3 @@ def create_tables() -> None:
     )"""
     execute(sql)
 
-def name_lookup(name: str) -> str:
-    if name == 'Kongming, Sleeping Dragon':
-        name = 'Kongming, "Sleeping Dragon"'
-    elif name == 'Pang Tong, Young Phoenix':
-        name = 'Pang Tong, "Young Phoenix"'
-    if not CARDS:
-        rs = db().execute(multiverse.base_query())
-        for row in rs:
-            CARDS[card.canonicalize(row['name'])] = row['name']
-    canonical = card.canonicalize(name)
-    if canonical not in CARDS:
-        raise InvalidDataException('Bogus name {name} ({canonical}) found.'.format(name=name, canonical=canonical))
-    return CARDS[canonical]
